Recalibrate FFT

+++ src/system/platform/pdm_handle.h
@@ -12,14 +12,14 @@ static constexpr int16_t gainUpdateBaseline = 80;
 
 struct PdmData
 {
-  static constexpr size_t SAMPLE_SIZE = 512; // same has samplesFFT;
+  static constexpr uint16_t SAMPLE_SIZE = 512;
   std::array<int16_t, SAMPLE_SIZE> data;
 
   uint64_t sampleDuration_us;
   uint64_t sampleTime_us;
 
   // number of sound sample in array
-  uint32_t sampleRead = 0;
+  uint16_t sampleRead = 0;
 
   bool is_valid() const { return sampleRead > 0; }
 };
